% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_lamy_mc_part.R
\name{fct_lamy_mc_part}
\alias{fct_lamy_mc_part}
\title{lamy_mc_part}
\usage{
fct_lamy_mc_part(Y, s, t)
}
\arguments{
\item{Y}{Community table. An observation by species matrix with row blocks corresponding to local patches
containing the biomass of species (i) in local patch (k).}

\item{s}{Number of patches.}

\item{t}{Number of years.}
}
\value{
Synchrony and stability metrics as spatial and organizational levels of the ecosystem.:
 
 \item{CV_SL}{Species-level variability at the local scale}
 \item{CV_CL}{Community-level variability at the local scale}
 \item{CV_SR}{Species-level variability at the regional scale; the weighted average of metapopulation variability across species}
 \item{CV_CR}{Community-level variability at the regional scale; the variability of the whole metacommunity}
 
Species-level table (i):

 \item{phi_i_LR}{Spatial synchrony of species i}
 \item{w_i}{Weight of species i (SD contribution to metacommunity SD)}

Patch-level table (k):

 \item{phi_SC_k}{Species synchrony within patch k}
 \item{w_k}{Weight of patch k (biomass contribution to metacommunity)}
}
\description{
Code provided in the supplement to Lamy et al. 2019, which has the following
description (lightly edited):

"Partitions ecological variability across two spatial scales (local and regional)
and two organizational levels (population and community).
Six synchrony metrics serve as scaling factors to measure how variability scales
from one specific hierarchical level to the next. These synchrony metrics are 
the square-root transformation of the synchrony metric developed by 
Loreau and de Mazancourt (2008)."
}
\references{
Wang S, Lamy T, Hallett LM and Loreau M (2019) Stability and synchrony across ecological hierarchies in heterogeneous metacommunities: linking theory to data.
Ecography. doi: 10.1111/ecog.04290.

Wang and Loreau (2014) Ecosystem stability in space: alpha, beta and gamma variability.
Ecology Letters. 17: 891-901.

Loreau M and de Mazancourt C (2008) Species synchrony and its drivers: neutral and nonneutral community dynamics in fluctuating environments.
American Naturalist. 172: E48-66.

Data S1 to Lamy T, Wang S, Renard D, Lafferty KD, Reed DC and Miller RJ (2019)
Species insurance trumps spatial insurance in stabilizing biomass of a marine macroalgal metacommunity. Ecology.
}
\author{
Thomas Lamy
}
