% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable.r
\name{Links}
\alias{Links}
\alias{loc_g}
\alias{loc_h}
\alias{disp_g}
\alias{disp_h}
\alias{skew_g}
\alias{skew_h}
\alias{tail_g}
\alias{tail_h}
\title{Links}
\usage{
loc_g(x)

loc_h(x)

disp_g(x)

disp_h(x)

skew_g(x)

skew_h(x)

tail_g(x)

tail_h(x)
}
\arguments{
\item{x}{the function argument}
}
\description{
Link and inverse functions for use in stablereg
}
