% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable-package.R
\docType{package}
\name{stable-package}
\alias{stable}
\alias{stable-package}
\title{stable: stable}
\description{
Applied synchrony-stability analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{github.com/seanhardison1/stable}
  \item Report bugs at \url{github.com/seanhardison1/stable/issues}
}

}
\author{
\strong{Maintainer}: Sean Hardison \email{sh5rs@virginia.edu}

}
\keyword{internal}
