% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_stable.R
\name{theme_stable}
\alias{theme_stable}
\title{A ggplot2 theme}
\usage{
theme_stable(...)
}
\description{
A ggplot2 theme
}
\examples{

library(ggplot2)
ggplot(iris) +
   geom_point(aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
   theme_stable()
}
