% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable.r
\name{Links}
\alias{Links}
\alias{disp.g}
\alias{disp.h}
\alias{loc.g}
\alias{loc.h}
\alias{skew.g}
\alias{skew.h}
\alias{tail.g}
\alias{tail.h}
\title{Links}
\usage{
loc.g(x)

loc.h(x)

disp.g(x)

disp.h(x)

skew.g(x)

skew.h(x)

tail.g(x)

tail.h(x)
}
\arguments{
\item{x}{the function argument}
}
\description{
Link and inverse functions for use in stablereg
}

